refactor ad implementation
