+++ ui/page/discover/view.jsx
@@ -13,7 +13,7 @@ import usePersistedState from 'effects/use-persisted-state';
 import analytics from 'analytics';
 import HiddenNsfw from 'component/common/hidden-nsfw';
 import Icon from 'component/common/icon';
-import Ads from 'web/component/ads';
+import Ads, { injectAd } from 'web/component/ads';
 import LbcSymbol from 'component/common/lbc-symbol';
 import I18nMessage from 'component/i18nMessage';
 import moment from 'moment';
@@ -198,115 +198,7 @@ function DiscoverPage(props: Props) {
       return;
     }
 
-    (async function () {
-      // test if adblock is enabled
-      let adBlockEnabled = false;
-      const googleAdUrl = 'https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js';
-      try {
-        await fetch(new Request(googleAdUrl)).catch((_) => {
-          adBlockEnabled = true;
-        });
-      } catch (e) {
-        adBlockEnabled = true;
-      } finally {
-        if (!adBlockEnabled) {
-          // select the cards on page
-          let cards = document.getElementsByClassName('card claim-preview--tile');
-
-          // eslint-disable-next-line no-inner-declarations
-          function checkFlag() {
-            if (cards.length === 0) {
-              window.setTimeout(checkFlag, 100);
-            } else {
-              // find the last fully visible card
-              let lastCard;
-
-              // width of browser window
-              const windowWidth = window.innerWidth;
-
-              // on small screens, grab the second item
-              if (windowWidth <= 900) {
-                lastCard = cards[1];
-              } else {
-                // otherwise, get the last fully visible card
-                for (const card of cards) {
-                  const isFullyVisible = isScrolledIntoView(card);
-                  if (!isFullyVisible) break;
-                  lastCard = card;
-                }
-              }
-
-              // clone the last card
-              // $FlowFixMe
-              const clonedCard = lastCard.cloneNode(true);
-
-              // insert cloned card
-              // $FlowFixMe
-              lastCard.parentNode.insertBefore(clonedCard, lastCard);
-
-              // delete last card so that it doesn't mess up formatting
-              // $FlowFixMe
-              // lastCard.remove();
-
-              // change the appearance of the cloned card
-              // $FlowFixMe
-              clonedCard.querySelector('.claim__menu-button').remove();
-
-              // $FlowFixMe
-              clonedCard.querySelector('.truncated-text').innerHTML = __(
-                'Hate these? Login to Odysee for an ad free experience'
-              );
-
-              // $FlowFixMe
-              clonedCard.querySelector('.claim-tile__info').remove();
-
-              // $FlowFixMe
-              clonedCard.querySelector('[role="none"]').removeAttribute('href');
-
-              // $FlowFixMe
-              clonedCard.querySelector('.claim-tile__header').firstChild.href = '/$/signin';
-
-              // $FlowFixMe
-              clonedCard.querySelector('.claim-tile__title').firstChild.removeAttribute('aria-label');
-
-              // $FlowFixMe
-              clonedCard.querySelector('.claim-tile__title').firstChild.removeAttribute('title');
-
-              // $FlowFixMe
-              clonedCard.querySelector('.claim-tile__header').firstChild.removeAttribute('aria-label');
-
-              // $FlowFixMe
-              clonedCard
-                .querySelector('.media__thumb')
-                .replaceWith(document.getElementsByClassName('homepageAdContainer')[0]);
-
-              // show the homepage ad which is not displayed at first
-              document.getElementsByClassName('homepageAdContainer')[0].style.display = 'block';
-
-              // $FlowFixMe
-              const imageHeight = window.getComputedStyle(lastCard.querySelector('.media__thumb')).height;
-              // $FlowFixMe
-              const imageWidth = window.getComputedStyle(lastCard.querySelector('.media__thumb')).width;
-
-              const styles = `#av-container, #AVcontent, #aniBox {
-                height: ${imageHeight} !important;
-                width: ${imageWidth} !important;
-              }`;
-
-              const styleSheet = document.createElement('style');
-              styleSheet.type = 'text/css';
-              styleSheet.id = 'customAniviewStyling';
-              styleSheet.innerText = styles;
-              // $FlowFixMe
-              document.head.appendChild(styleSheet);
-
-              window.dispatchEvent(new CustomEvent('scroll'));
-            }
-          }
-          checkFlag();
-        }
-      }
-    })();
+    injectAd();
   }, [isAuthenticated]);
 
   // Sync liveSection --> liveSectionStore
