+++ ui/page/home/view.jsx
@@ -1,8 +1,8 @@
 // @flow
 import * as ICONS from 'constants/icons';
 import * as PAGES from 'constants/pages';
-import { SHOW_ADS, SITE_NAME, SIMPLE_SITE, ENABLE_NO_SOURCE_CLAIMS } from 'config';
-import Ads from 'web/component/ads';
+import { SITE_NAME, SIMPLE_SITE, ENABLE_NO_SOURCE_CLAIMS } from 'config';
+import Ads, { injectAd } from 'web/component/ads';
 import React from 'react';
 import Page from 'component/page';
 import Button from 'component/button';
@@ -132,119 +132,7 @@ function HomePage(props: Props) {
   }
 
   React.useEffect(() => {
-    if (authenticated || !SHOW_ADS) {
-      return;
-    }
-
-    (async () => {
-      // test if adblock is enabled
-      let adBlockEnabled = false;
-      const googleAdUrl = 'https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js';
-      try {
-        await fetch(new Request(googleAdUrl)).catch((_) => {
-          adBlockEnabled = true;
-        });
-      } catch (e) {
-        adBlockEnabled = true;
-      } finally {
-        if (!adBlockEnabled) {
-          // select the cards on page
-          let cards = document.getElementsByClassName('card claim-preview--tile');
-
-          // eslint-disable-next-line no-inner-declarations
-          function checkFlag() {
-            if (cards.length === 0) {
-              window.setTimeout(checkFlag, 100);
-            } else {
-              // find the last fully visible card
-              let lastCard;
-
-              // width of browser window
-              const windowWidth = window.innerWidth;
-
-              // on small screens, grab the second item
-              if (windowWidth <= 900) {
-                lastCard = cards[1];
-              } else {
-                // otherwise, get the last fully visible card
-                for (const card of cards) {
-                  const isFullyVisible = isScrolledIntoView(card);
-                  if (!isFullyVisible) break;
-                  lastCard = card;
-                }
-
-                // if no last card was found, just exit the function to not cause errors
-                if (!lastCard) return;
-              }
-
-              // clone the last card
-              // $FlowFixMe
-              const clonedCard = lastCard.cloneNode(true);
-
-              // insert cloned card
-              // $FlowFixMe
-              lastCard.parentNode.insertBefore(clonedCard, lastCard);
-
-              // change the appearance of the cloned card
-              // $FlowFixMe
-              clonedCard.querySelector('.claim__menu-button').remove();
-
-              // $FlowFixMe
-              clonedCard.querySelector('.truncated-text').innerHTML = __(
-                'Hate these? Login to Odysee for an ad free experience'
-              );
-
-              // $FlowFixMe
-              clonedCard.querySelector('.claim-tile__info').remove();
-
-              // $FlowFixMe
-              clonedCard.querySelector('[role="none"]').removeAttribute('href');
-
-              // $FlowFixMe
-              clonedCard.querySelector('.claim-tile__header').firstChild.href = '/$/signin';
-
-              // $FlowFixMe
-              clonedCard.querySelector('.claim-tile__title').firstChild.removeAttribute('aria-label');
-
-              // $FlowFixMe
-              clonedCard.querySelector('.claim-tile__title').firstChild.removeAttribute('title');
-
-              // $FlowFixMe
-              clonedCard.querySelector('.claim-tile__header').firstChild.removeAttribute('aria-label');
-
-              // $FlowFixMe
-              clonedCard
-                .querySelector('.media__thumb')
-                .replaceWith(document.getElementsByClassName('homepageAdContainer')[0]);
-
-              // show the homepage ad which is not displayed at first
-              document.getElementsByClassName('homepageAdContainer')[0].style.display = 'block';
-
-              const thumbnail = window.getComputedStyle(lastCard.querySelector('.media__thumb'));
-
-              const styles = `#av-container, #AVcontent, #aniBox {
-                height: ${thumbnail.height} !important;
-                width: ${thumbnail.width} !important;
-              }`;
-
-              const styleSheet = document.createElement('style');
-              styleSheet.type = 'text/css';
-              styleSheet.id = 'customAniviewStyling';
-              styleSheet.innerText = styles;
-              // $FlowFixMe
-              document.head.appendChild(styleSheet);
-
-              // delete last card to not introduce layout shifts
-              lastCard.remove();
-
-              // addresses bug where ad doesn't show up until a scroll event
-              document.dispatchEvent(new CustomEvent('scroll'));
-            }
-          }
-          checkFlag();
-        }
-      }
-    })();
+    injectAd()
   }, []);
 
   return (
