+++ web/component/ads/index.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux';
 import { selectTheme } from 'redux/selectors/settings';
 import { makeSelectClaimForUri, selectClaimIsNsfwForUri } from 'redux/selectors/claims';
 import { selectUserVerifiedEmail } from 'redux/selectors/user';
-import Ads from './view';
+import Ads, { injectAd } from './view';
 
 const select = (state, props) => ({
   theme: selectTheme(state),
@@ -12,3 +12,4 @@ const select = (state, props) => ({
 });
 
 export default connect(select)(Ads);
+export { injectAd };
